+++ php/controller/register_cpf.php
@@ -6,6 +6,7 @@
   
   $data = json_decode(file_get_contents("php://input"));
   $name = @$data -> name;
+  $nickname = @$data -> nickname;
   $email = @$data -> email;
   $password = @$data -> password;
   $confirm_password = @$data -> confirm_password;
@@ -14,13 +15,14 @@
   $data_nasc = @$data -> data_nasc;
   $cep = @$data -> cep;
   $street = @$data -> street;
+  $street_number = @$data -> street_number;
   $city = @$data -> city;
   $state = @$data -> state;
   $tel = @$data -> tel;
   $cel = @$data -> cel;
   $code = @$data -> code;
 
-  if (empty($name) && empty($email) && empty($password) && empty($confirm_password) && empty($cpf) && empty($data_nasc) && empty($cep) && empty($street) && empty($city) && empty($state) && empty($cel) && empty($code)) {
+  if (empty($name) && empty($email) && empty($password) && empty($confirm_password) && empty($cpf) && empty($data_nasc) && empty($cep) && empty($street) && empty($city) && empty($state) && empty($cel) && empty($code) && empty($nickname) && empty($street_number)) {
     echo json_encode(array("status" => "error", "message" => true, "text" => "Preencha todos os campos"));
   }
   else if ($password != $confirm_password) {
@@ -63,8 +65,8 @@
         echo json_encode(array("status" => "error", "message" => true, "text" => "CPF já cadastrado"));
       }
       else {
-        $password = password_hash($password, PASSWORD_DEFAULT);
-        $sql = "INSERT INTO users (name, email, password, cpf, nasc, cep, street, city, state, tel, cel) VALUES ('$name', '$email', '$password', '$cpf', '$data_nasc', '$cep', '$street', '$city', '$state', '$tel', '$cel')";
+        $password = md5($password);
+        $sql = "INSERT INTO users (name, nickname, email, password, cpf, nasc, cep, street, street_number, city, state, tel, cel) VALUES ('$name', '$nickname', '$email', '$password', '$cpf', '$data_nasc', '$cep', '$street', '$street_number', '$city', '$state', '$tel', '$cel')";
         $result = mysqli_query($con, $sql);
         if ($result) {
           echo json_encode(array("status" => "success", "message" => false, "text" => "Cadastro realizado com sucesso"));
